% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeScripts.R
\name{classing}
\alias{classing}
\title{Class and Save Object}
\usage{
classing(x, dir, name)
}
\arguments{
\item{x}{The object to be saved.}

\item{dir}{The directory where the object will be saved.}

\item{name}{The name of the file (without extension).}
}
\value{
None
}
\description{
This function saves an object based on its class.
}
