% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeScripts.R
\name{finalD}
\alias{finalD}
\title{Finalize a data file}
\usage{
finalD(name)
}
\arguments{
\item{name}{The name of the data file (with extension).}
}
\value{
None
}
\description{
This function finalizes a data file by copying it to the final and backup directories.
}
