% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeScripts.R
\name{finalLib}
\alias{finalLib}
\title{Finalize a library script}
\usage{
finalLib(name)
}
\arguments{
\item{name}{The name of the library script file (without extension).}
}
\value{
None
}
\description{
This function finalizes a library script by copying it to the final and backup directories.
}
