% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeScripts.R
\name{finalScript}
\alias{finalScript}
\title{Finalize a Script File}
\usage{
finalScript(name)
}
\arguments{
\item{name}{The name of the script file (without extension).}
}
\value{
None
}
\description{
This function finalizes a script file by copying it to the final and backup directories.
}
