% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeScripts.R
\name{newProject}
\alias{newProject}
\title{Create a New Project Directory Structure}
\usage{
newProject(dir, projectName)
}
\arguments{
\item{dir}{The base directory where the project will be created.}

\item{projectName}{The name of the project.}
}
\value{
A message indicating the status of the project creation.
}
\description{
This function creates a new project directory with a specified structure.
}
