% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeScripts.R
\name{processLib}
\alias{processLib}
\title{Process a common library script file.}
\usage{
processLib(name)
}
\arguments{
\item{name}{The name of the library script file (without extension).}
}
\value{
None
}
\description{
This function processes a common library script file.
}
