% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeScripts.R
\name{processScript}
\alias{processScript}
\title{Process a Script File}
\usage{
processScript(name)
}
\arguments{
\item{name}{The name of the script file (without extension).}
}
\value{
None
}
\description{
This function processes a script file within the project directory.
}
