% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeScripts.R
\name{runSource}
\alias{runSource}
\title{Run the source script}
\usage{
runSource(name)
}
\arguments{
\item{name}{The name of the source script (without extension).}
}
\value{
None
}
\description{
This function runs the script name in the source directory.
}
