% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeScripts.R
\name{saveD}
\alias{saveD}
\title{Save a data file}
\usage{
saveD(x, name)
}
\arguments{
\item{x}{The object to be saved.}

\item{name}{The name of the file (without extension).}
}
\value{
None
}
\description{
This function saves a data file to the folder.
}
