% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeScripts.R
\name{saveF}
\alias{saveF}
\title{Save Final Object}
\usage{
saveF(x, dir, name)
}
\arguments{
\item{x}{The object to be saved.}

\item{dir}{The subdirectory within the project where the object will be saved.}

\item{name}{The name of the file (without extension).}
}
\value{
None
}
\description{
This function saves an object to a specified directory within the project.
}
