% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeScripts.R
\name{saveSource}
\alias{saveSource}
\title{Create/edit a source script in the source directory}
\usage{
saveSource(name)
}
\arguments{
\item{name}{The name of the source script (without extension).}
}
\value{
None
}
\description{
This function creates or edits a source script in the source directory.
}
