% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeScripts.R
\name{setCommonLib}
\alias{setCommonLib}
\title{Set or get common R library}
\usage{
setCommonLib(dir = "", overwrite = F)
}
\arguments{
\item{dir}{The directory for the common library.}

\item{overwrite}{Do you want to overwrite the current directory?}
}
\value{
None
}
\description{
This function sets the common R library or gets it when it already exists.
}
