% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeScripts.R
\name{setProject}
\alias{setProject}
\title{Set Project Directory}
\usage{
setProject(projectName)
}
\arguments{
\item{projectName}{The name of the project.}
}
\value{
None
}
\description{
This function sets the default project directory.
}
