% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeScripts.R
\name{tempF}
\alias{tempF}
\title{Save or import temporary files}
\usage{
tempF(x = NULL, name)
}
\arguments{
\item{x}{The object to be saved. Ignore for importing data.}

\item{name}{Saving: the name of the file (without extension) and
importing: the name of the file (with extension)}
}
\value{
None
}
\description{
This function saves an object to the temporary data directory or 
imports a temporary file into the global environment.
}
