% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeScripts.R
\name{classingImport}
\alias{classingImport}
\title{Class and import Object}
\usage{
classingImport(dir)
}
\arguments{
\item{dir}{The directory where the object is saved.}

\item{name}{The name of the file (with extension).}
}
\value{
None
}
\description{
This function imports an object based on its class.
}
