% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizeScripts.R
\name{importD}
\alias{importD}
\title{Import a data file}
\usage{
importD(name)
}
\arguments{
\item{name}{The name of the data file (with extension).}
}
\value{
None
}
\description{
This function opens a datafile
}
